Refactor: Update example server and README
Update comprehensive example server and README to reflect current best practices and structure.
